pyproject.toml
[tool.poetry]
name = "medscrape"
version = "0.0.1"
description = "webscraper for medical school content"
authors = ["Jay Scambler <jay@greyhaven.ai>"]
license = "MIT"
readme = "README.md"
packages = [{include = "medscrape"}]
repository = "https://github.com/greyhaven-ai/medscrape"

[tool.poetry.dependencies]
asyncio = "3.4.3"
aiolimiter = "1.1.0"
bandit = "1.7.8"
beautifulsoup4 = "4.11.2" 
fastapi = "0.110.0"
instructor = ">=0.5.2"
lancedb = ">=0.6.3"
openai = "1.13.3"
psutil = "5.9.4"
pydantic = "2.6.3"
python = ">=3.11,<3.12"
python-dotenv = "1.0.1"
redis = "4.5.4"
requests = "2.31.0"
retry = "^0.9.2"
sentence-transformers = "2.5.1"
tantivy = "0.21.0"
uvicorn = "0.28.0"
unstructured = "0.12.5"


[tool.poetry.scripts]
medscrape = "medscrape.cli.cli:app"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"
mkdocstrings = "^0.23.0"
mkdocstrings-python = "^1.8.0"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.12"
mkdocs-material-extensions = "^1.3.1"
mypy = "^1.8.0"
poetry = "^1.8.2"
pymdown-extensions = "^10.7.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
ipykernel = "^6.29.3"


[tool.poetry.group.test-docs.dependencies]
mkdocstrings = "0.23.0"
mkdocstrings-python = "1.8.0"
mkdocs = "^1.5.3"
mkdocs-material = "9.5.12"
mkdocs-material-extensions = "1.3.1"
pymdown-extensions = "10.7.0"

[tool.poetry.group.test.dependencies]
coverage = "^7.3.2"
mypy = "^1.8.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
